feat: 통계 조회 DTO 추가
- 통계 조회와 관련된 DTO 객체를 추가했습니다.

Resolves: #31

+++ backend/src/main/java/com/ssafy/a303/backend/user/dto/GetStatisticsResponseDto.java
@@ -1,9 +1,9 @@
 package com.ssafy.a303.backend.user.dto;
 
 public record GetStatisticsResponseDto(
-        int totalDaysCount,
-        int streakDaysCount,
-        int monthDaysCount,
-        int totalWordsCount,
-        int totalFoldersCount
+        long totalDaysCount,
+        long streakDaysCount,
+        long monthDaysCount,
+        long totalWordsCount,
+        long totalFoldersCount
 ) {}
